rename avx to avx2, compile on android
